+++ server/src/bruFiles/completions/mapVariablesToCompletions.ts
@@ -1,9 +1,9 @@
 import {
     VariableReferenceType,
     Range,
-    Position,
     VariableAvailabilityScopes,
     RequestFileBlockName,
+    LineBreakType,
 } from "@global_shared";
 import {
     VariableSpecificRequestData,
@@ -21,6 +21,12 @@ import {
     MatchingDynamicVariables,
 } from "../shared/interfaces";
 
+export type GetTextEditForVariableCompletion = (
+    variableName: string,
+    rangeToReplace: Range,
+    lineBreak: LineBreakType,
+) => TextEdit;
+
 export function mapVariablesToCompletions(
     matchingReferences: {
         staticEnvVariables: {
@@ -32,7 +38,7 @@ export function mapVariablesToCompletions(
         dynamicVariables: MatchingDynamicVariables;
     },
     requestData: VariableSpecificRequestData,
-    appendOnInsertion?: string,
+    getTextEditForCompletion?: GetTextEditForVariableCompletion,
 ) {
     const { staticEnvVariables, staticScriptVariables, dynamicVariables } =
         matchingReferences;
@@ -41,7 +47,7 @@ export function mapVariablesToCompletions(
         dynamicVariables,
         {
             prefixForSortText: "a",
-            appendOnInsertion,
+            getTextEditForCompletion,
         },
     );
 
@@ -49,7 +55,7 @@ export function mapVariablesToCompletions(
         ? mapStaticScriptVariables(requestData, staticScriptVariables, {
               // Display static script variables below dynamic ones, but above static environment variables.
               prefixForSortText: "b",
-              appendOnInsertion,
+              getTextEditForCompletion,
           })
         : [];
 
@@ -60,7 +66,10 @@ export function mapVariablesToCompletions(
             dynamicVariables,
         ),
         // Display static environment variables below dynamic ones and static script variables.
-        { prefixForSortText: "c", appendOnInsertion },
+        {
+            prefixForSortText: "c",
+            getTextEditForCompletion,
+        },
     );
 
     return resultsForDynamicVariables.concat(
@@ -73,13 +82,16 @@ function mapStaticScriptVariables(
     {
         variable: { start, end },
         functionType: sourceReferenceType,
+        documentLineBreak,
     }: VariableSpecificRequestData,
     allReferences: EquivalentVariableReferencesFromOtherFiles[],
     modifications: {
         prefixForSortText: string;
-        appendOnInsertion?: string;
+        getTextEditForCompletion?: GetTextEditForVariableCompletion;
     },
 ): CompletionItem[] {
+    const { prefixForSortText, getTextEditForCompletion } = modifications;
+
     return allReferences.map(
         ({
             mostRelevantReference: {
@@ -110,16 +122,17 @@ function mapStaticScriptVariables(
                     ? `WARNING: Will overwrite static script variable.`
                     : undefined,
                 sortText: getSortText(
-                    modifications.prefixForSortText,
+                    prefixForSortText,
                     variableName,
                     indirectionLevel,
                 ),
-                textEdit: getTextEdit(
-                    variableName,
-                    start,
-                    end,
-                    modifications.appendOnInsertion,
-                ),
+                textEdit: getTextEditForCompletion
+                    ? getTextEditForCompletion(
+                          variableName,
+                          new Range(start, end),
+                          documentLineBreak,
+                      )
+                    : undefined,
             };
         },
     );
@@ -130,7 +143,7 @@ function mapDynamicVariables(
     { fromSameFile, fromOtherFiles }: MatchingDynamicVariables,
     modifications: {
         prefixForSortText: string;
-        appendOnInsertion?: string;
+        getTextEditForCompletion?: GetTextEditForVariableCompletion;
     },
 ) {
     const groupedRefs = groupReferencesByName({ fromSameFile, fromOtherFiles });
@@ -168,7 +181,10 @@ function mapDynamicVariables(
 }
 
 function getCompletionForRefsWithinOwnFile(
-    { variable: { start, end } }: VariableSpecificRequestData,
+    {
+        variable: { start, end },
+        documentLineBreak,
+    }: VariableSpecificRequestData,
     groupedReferences: {
         variableName: string;
         referenceType: VariableReferenceType;
@@ -177,9 +193,10 @@ function getCompletionForRefsWithinOwnFile(
     },
     modifications: {
         prefixForSortText: string;
-        appendOnInsertion?: string;
+        getTextEditForCompletion?: GetTextEditForVariableCompletion;
     },
 ): CompletionItem {
+    const { prefixForSortText, getTextEditForCompletion } = modifications;
     const {
         detailsForOwnFileRefs: {
             blockName,
@@ -214,23 +231,22 @@ function getCompletionForRefsWithinOwnFile(
                           ? ""
                           : ` and in ${referencesFromOtherFiles.otherMatchingReferences.length + 1} other file(s)`,
                   ),
-        sortText: getSortText(
-            modifications.prefixForSortText,
-            variableName,
-            0,
-            blockName,
-        ),
-        textEdit: getTextEdit(
-            variableName,
-            start,
-            end,
-            modifications.appendOnInsertion,
-        ),
+        sortText: getSortText(prefixForSortText, variableName, 0, blockName),
+        textEdit: getTextEditForCompletion
+            ? getTextEditForCompletion(
+                  variableName,
+                  new Range(start, end),
+                  documentLineBreak,
+              )
+            : undefined,
     };
 }
 
 function getCompletionForRefsFromOnlyOtherFiles(
-    { variable: { start, end } }: VariableSpecificRequestData,
+    {
+        variable: { start, end },
+        documentLineBreak,
+    }: VariableSpecificRequestData,
     groupedReferences: {
         variableName: string;
         referenceType: VariableReferenceType;
@@ -238,7 +254,7 @@ function getCompletionForRefsFromOnlyOtherFiles(
     },
     modifications: {
         prefixForSortText: string;
-        appendOnInsertion?: string;
+        getTextEditForCompletion?: GetTextEditForVariableCompletion;
     },
 ): CompletionItem {
     const {
@@ -249,6 +265,7 @@ function getCompletionForRefsFromOnlyOtherFiles(
             otherMatchingReferences,
         },
     } = groupedReferences;
+    const { prefixForSortText, getTextEditForCompletion } = modifications;
 
     return {
         label: variableName,
@@ -261,16 +278,17 @@ function getCompletionForRefsFromOnlyOtherFiles(
                 ? undefined
                 : `Relevant reference(s) in ${otherMatchingReferences.length} other file(s).`,
         sortText: getSortText(
-            modifications.prefixForSortText,
+            prefixForSortText,
             variableName,
             mostRelevantReference.indirectionLevel,
         ),
-        textEdit: getTextEdit(
-            variableName,
-            start,
-            end,
-            modifications.appendOnInsertion,
-        ),
+        textEdit: getTextEditForCompletion
+            ? getTextEditForCompletion(
+                  variableName,
+                  new Range(start, end),
+                  documentLineBreak,
+              )
+            : undefined,
     };
 }
 
@@ -320,18 +338,6 @@ function getKind(referenceType: VariableReferenceType) {
     return referenceType == VariableReferenceType.Read
         ? CompletionItemKind.Field
         : CompletionItemKind.Function;
-}
-
-function getTextEdit(
-    variableName: string,
-    start: Position,
-    end: Position,
-    appendOnInsertion?: string,
-): TextEdit {
-    return {
-        newText: `${variableName}${appendOnInsertion ?? ""}`,
-        range: new Range(start, end),
-    };
 }
 
 function getSortText(