+++ shared/languageUtils/commonBlocks/generic/variables/interfaces.ts
@@ -1,4 +1,0 @@
-export enum VariableNameMatchingMode {
-    Exact = 1,
-    Ignore = 2,
-}